refactor: split server code into separate single purpose files
