+++ package.json
@@ -31,14 +31,11 @@
   "homepage": "http://dat-data.com",
   "dependencies": {
     "accountdown": "^3.0.0",
-    "body": "^4.5.0",
     "bootstrap": "^3.3.1",
-    "bytewise": "^0.7.1",
     "changes-feed": "^1.0.0",
     "changesdown": "^2.1.0",
     "concat-stream": "^1.4.7",
     "cookie-auth": "^2.4.0",
-    "corsify": "^2.0.0",
     "dat-api-client": "^1.0.0",
     "dathub-api-client": "^1.1.0",
     "debug": "^2.0.0",
@@ -48,7 +45,6 @@
     "is-my-json-valid": "^1.4.0",
     "is-url": "^1.1.0",
     "jquery": "^2.1.1",
-    "level-indexer": "^2.0.0",
     "level-prebuilt": "^0.18.0",
     "nodemon": "^1.2.1",
     "page": "^1.4.1",
@@ -58,10 +54,8 @@
     "response": "^0.14.0",
     "rest-parser": "^1.0.5",
     "routes-router": "^4.1.1",
-    "sqlite-search": "^3.2.0",
     "st": "^0.5.2",
-    "subleveldown": "^1.0.6",
-    "xhr": "^1.17.0"
+    "subleveldown": "^1.0.6"
   },
   "devDependencies": {
     "autoless": "^0.1.5",
